requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
redis==5.0.1
python-socketio==5.10.0
gunicorn==21.2.0
gevent==24.2.1
gevent-websocket==0.10.1
python-dotenv==1.0.0

